Changed x,y coordinates from long to int

+++ src/main/java/core/Grid.java
@@ -86,7 +86,7 @@ public class Grid {
      * @param y Współrzędna y komórki
      * @return Stan komórki typu {@link core.Cell.State}. Jeśli komórka nie jest znaleziona zwraca stan EMPTY
      */
-    public Cell.State getCellState(long x, long y){
+    public Cell.State getCellState(int x, int y){
         Coordinate coord = new Coordinate(x,y);
         if (this.hashMap.containsKey(coord)){
             return hashMap.get(coord).getState();
